requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
speedtest-cli==2.1.3
pandas==2.0.3
numpy==1.24.3
gunicorn==21.2.0